[refactoring] step2->lotto 패키지명 변경
