+++ src/main/java/lotto/domain/Lotto.java
@@ -1,9 +1,7 @@
-package step2.domain;
+package lotto.domain;
 
-import step2.domain.starategy.GetLottoNumberStrategy;
+import lotto.domain.starategy.GetLottoNumberStrategy;
 
-import java.util.ArrayList;
-import java.util.Collections;
 import java.util.List;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
