+++ src/main/java/lotto/domain/Prize.java
@@ -1,4 +1,4 @@
-package step2.domain;
+package lotto.domain;
 
 import java.util.Arrays;
 import java.util.List;
