+++ src/main/java/lotto/view/Input.java
@@ -1,4 +1,4 @@
-package step2.view;
+package lotto.view;
 
 import java.util.Arrays;
 import java.util.Scanner;
